git_access: Add --git-noclean option
This option can be used when local commits should be kept. This is for
example useful to add images, commit them, run the script and have it
updated the addon.xml files accordingly.

If specified:
- On clone: don't reset, but rebase
- On commit: squash commits

Also init new GitHub repos so that they get an initial empty comment
and we don't have to write extra code for dealing with empty repos.