template_processor: Fix script after adding language attributes
When parsing the existing addon.xml.in files, xmljson created a
shortened output if the element didn't contain any attributes.

This behavior broke the script with 2b9073f where attributes were added
and the element content has to be accessed with '.content' now.

Disable the behavior in xmljson and always for the creation of
'content' fields so that we always access the content the same way
(no matter if there are attributes or not).