process_game_addons: Determine assets only when generating addon files
When running the script with an empty working directory, the generated
addon.xml.in files always have the assets removed.

This is because the asset list is compiled before the source repos
are cloned.

This change moves finding the assets to the generation step where
we are sure that we have the repo checked out.