templates: Simplify templates
Simplify templates by restructuring some of the info values.
The addon name (for example game.libretro.2048) is needed in many
places, so far we used `game.libretro.{{ game.name }}` in the
templates, now we use `game.addon`.

In the ADDONS dict the options dict is no more optional, this
simplifies setting up the info structure in `process_game_addons.py`.